Attempt to fix rotation difficulties and alignment issues.
Tried to fix the rotation difficulties of the planet gears. Decreased the planet gear module by a bit to add tolerance. Tried to fix alignment issues due to too small carrier tolerances. Increased tolerances of carrier bars to the inner of the hole of the planet gears. Also increased carrier bar thickness to strengthen the part - from now on it should be printed with 100% infill.